llcppsymg/refactor: common pkg to tool/
