+++ _xtool/llcppsymg/tool/config/config.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/goplus/lib/c"
 	"github.com/goplus/lib/c/clang"
-	"github.com/goplus/llcppg/_xtool/llcppsymg/clangutils"
+	clangutils "github.com/goplus/llcppg/_xtool/llcppsymg/tool/clang"
 	llcppg "github.com/goplus/llcppg/config"
 	llgoc "github.com/goplus/llgo/c"
 	"github.com/goplus/llpkg/cjson"
